requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pydicom==3.0.1
pytest==8.4.1
tensorflow==2.20.0
tkcap==0.0.4






